Move utils outside transformers folder
