+++ flip/__init__.py
@@ -5,3 +5,4 @@ flip
 __version__ = "0.0.1"
 
 from . import transformers
+from . import utils
